exploring_mars v0.1.0 ExploringMars.Mission

This module defines a mission. A mission encapsulates a single run of a probe, being defined by a single set of problem parameters (bounds, initial position, and instructions).

Calling ExploringMars.Mission.run/3 with these parameters will produce the status in which the probe ended (mission outcome -- see below), together with the final position of the probe.

This module should change if the way a mission is represented changes.

Types

outcome()

outcome() :: :ok | :out_of_bounds | :illegal_instruction

A mission outcome represents the final status of a mission.

:ok → All commands were executed successfully

:out_of_bounds → The probe fell off the plateau. The position reported will be the position the probe was attempting to move to when it fell off.

:illegal_instruction → The probe tried to execute something that was not a valid instruction. The position reported will be the position where the probe was when it executed the invalid instruction.

result()

result() :: {outcome(), ExploringMars.Mission.Position.t()}

A mission's result is its outcome together with its final position. See the outcome/0 type's documentation above for further details.

Functions

result_to_string(mission_result)

result_to_string(result()) :: String.t()

Converts the result of a mission into a string suitable for user-facing output.

Examples

iex> Mission.result_to_string({:ok, {{1, 1}, :N}})
"1 1 N\n"

iex> Mission.result_to_string({:out_of_bounds, {{1, 1}, :N}})
"OUT OF BOUNDS @ 1 1 N\n"

iex> Mission.result_to_string({:invalid_instruction, {{1, 1}, :N}})
"INVALID INSTRUCTION @ 1 1 N\n"

run(bounds, position, instructions)

run(ExploringMars.Mission.Coordinate.t(), ExploringMars.Mission.Position.t(), [
  ExploringMars.Mission.Instruction.t()
]) :: result()

Run a mission by setting its parameters. Returns mission outcome and final position. Each outcome specifies how its final position should be interpreted. Check the outcome/0 type's documentation above for further details.

Examples

iex> Mission.run({2, 2}, {{0, 0}, :N}, [:M, :R, :M])
{:ok, {{1, 1}, :E}}

iex> Mission.run({1, 1}, {{0, 0}, :N}, [:M, :R, :M, :M])
{:out_of_bounds, {{2, 1}, :E}}

iex> Mission.run({1, 1}, {{0, 0}, :N}, [:M, :R, "%", :M, :M])
{:invalid_instruction, {{0, 1}, :E}}