exploring_mars v0.1.0 ExploringMars.Mission.Coordinate

This module defines functions that create and operate on coordinates in the probe's space of movement.

This module and the Direction module should change if the coordinate representation used in the problem changes - for instance, if we choose to use {y, x} coordinates instead of {x, y} or if we choose to model a probe moving in 3D space.

This means this module is coupled to the Direction model, as it depends directly on the representation of the Direction.t type. This is acceptable for now, as the Direction.t type is quite simple, but it might need change. Check the note in the documentation of the Direction module for further details.

Types

t()

t() :: {integer(), integer()}

A coordinate is defined as a pair of integers specifying a position on the cartesian plane. The order of coordinates is {x, y}.

Functions

from_strings(x_string, y_string)

from_strings(String.t(), String.t()) :: {:ok, t()} | {:no_parse, String.t()}

Takes a pair of strings, x_string and y_string and tries to parse them as a coordinate.

Examples

iex> Coordinate.from_strings("100", "200") {:ok, {100, 200}} iex> Coordinate.from_strings("100x", "200y") {:no_parse, "{100x, 200y}"}

move(coordinate, direction)

move(t(), ExploringMars.Mission.Direction.t()) :: t()

Moves a coordinate along a direction.

Examples

iex> Coordinate.move({-1, 0}, :E) {0, 0} iex> Coordinate.move({0, 0}, "Not a direction") ** (ArgumentError) argument is not a direction

pretty_print(coordinate)

pretty_print(t()) :: String.t()

Converts a coordinate into a representation suitable for user-facing output.