exploring_mars v0.1.0 ExploringMars.Mission.Instruction

This module defines functions that operate on the probe's instructions.

This module - and only this module - should change if new instructions are added, instructions are removed or if the meaning of instructions is changed.

Types

t()

t() :: :L | :R | :M

A single instruction, one of :L, :R, :M, representing "Turn left", "Turn right", and "Move forward", respectively. Use as Instruction.t.

Functions

from_string(string)

from_string(String.t()) :: t() | {:no_parse, String.t()}

Takes a string and tries to parse it as an instruction.

Examples

iex> Instruction.from_string("L")
:L

iex> Instruction.from_string("")
{:no_parse, ""}

run(position, instruction)

run(ExploringMars.Mission.Position.t(), t()) ::
  ExploringMars.Mission.Position.t() | :invalid_instruction

Takes a probe position and an instruction and returns the next position of the probe.

Examples

iex> Instruction.run({{0, 0}, :N}, :M)
{{0, 1}, :N}

iex> Instruction.run({{0, 0}, :N}, :L)
{{0, 0}, :W}

iex> Instruction.run({{0, 0}, :N}, :K)
:invalid_instruction