exploring_mars v0.1.0 ExploringMars.Mission.Position

This module defines functions that operate on the probe's position.

This module should change if the way a position is specified changes. For instance, if we had 3D space instead of 2D, or further degrees of freedom. See the notes in the Coordinate and Direction modules.

Types

t()

t() ::
  {ExploringMars.Mission.Coordinate.t(), ExploringMars.Mission.Direction.t()}

The probe's position, specified by an x-coordinate, y-coordinate and direction. Use as Position.t.

Functions

from_strings(x_string, y_string, direction_string)

from_strings(String.t(), String.t(), String.t()) ::
  {:ok, t()} | {:no_parse, String.t()}

Takes three strings representing the x_coordinate, y_coordinate and direction of the probe and returns either :ok and the specified position or :no_parse and a description of what went wrong.

Examples

iex> Position.from_strings("1", "1", "N")
{:ok, {{1, 1}, :N}}

iex> Position.from_strings("1", "1", "K")
{:no_parse, "Invalid direction K"}

iex> Position.from_strings("x", "1", "N")
{:no_parse, "Invalid coordinates {x, 1}"}

iex> Position.from_strings("x", "1", "K")
{:no_parse, "Invalid coordinates {x, 1}, Invalid direction K"}

move_forward(position)

move_forward(t()) :: t()

Takes a position and returns the position obtained by moving forward from that position.

Examples

iex> Position.move_forward({{0, 0}, :N})
{{0, 1}, :N}

iex> Position.move_forward({{0, 0}, :W})
{{-1, 0}, :W}

pretty_print(position)

pretty_print(t()) :: String.t()

Converts a position into a representation suitable for user-facing output.

Examples

iex> Position.pretty_print({{2, 2}, :S})
"2 2 S"

turn_left(position)

turn_left(t()) :: t()

Takes a position and returns the position obtained by turning left from that position.

Examples

iex> Position.turn_left({{2, 2}, :N})
{{2, 2}, :W}

turn_right(position)

turn_right(t()) :: t()

Takes a position and returns the position obtained by turning right from that position.

Examples

iex> Position.turn_right({{2, 2}, :N})
{{2, 2}, :E}