exploring_mars v0.1.0 ExploringMars.MissionRunner

File I/O handling module. Reads parameters from an io_device (which should be either a file or :stdio and runs each mission using the Mission module.

This module should change if the specification of how mission parameters are laid out changes.

Functions

get_bounds_and_run(input, output)

get_bounds_and_run(File.io_device(), File.io_device()) :: :ok

Takes an input file, an output file, and tries to read the input file as a mission specification, containing bounds and several missions.

Outputs for each mission either an error (if the mission was malformed) or a mission result. For more information on mission results, check the documentation for the Mission module.